+++ docs/source/user/combinators.rst
@@ -247,78 +247,3 @@ is a simple sample of individuals from a theoretical larger dataset that capture
 
 The efficacy of this combinator improves as n gets larger.
 
-
-User-Defined Callbacks
-----------------------
-
-It is possible to construct Transformations, Measurements and Postprocessors on your own via Python functions.
-
-.. list-table::
-   :header-rows: 1
-
-   * - Component
-     - Constructor
-   * - Transformation
-     - :func:`opendp.combinators.make_user_transformation`
-   * - Measurement
-     - :func:`opendp.combinators.make_user_measurement`
-   * - Postprocessor
-     - :func:`opendp.combinators.make_user_postprocessor`
-
-.. note::
-
-    This requires a looser trust model, as we cannot verify any privacy or stability properties of user-defined functions.
-
-    .. doctest::
-
-        >>> enable_features("honest-but-curious")
-
-In this example, we mock the typical API of the OpenDP library:
-
-.. doctest::
-
-    >>> from opendp.combinators import make_user_transformation
-    >>> from opendp.domains import vector_domain, atom_domain
-    >>> from opendp.metrics import symmetric_distance
-    >>> from opendp.typing import *
-    ...
-    >>> def make_repeat(multiplicity):
-    ...     """Constructs a Transformation that duplicates each record `multiplicity` times"""
-    ...     def function(arg: List[int]) -> List[int]:
-    ...         return arg * multiplicity
-    ... 
-    ...     def stability_map(d_in: int) -> int:
-    ...         # if a user could influence at most `d_in` records before, 
-    ...         # they can now influence `d_in` * `multiplicity` records
-    ...         return d_in * multiplicity
-    ...
-    ...     return make_user_transformation(
-    ...         vector_domain(atom_domain(T=int)),
-    ...         vector_domain(atom_domain(T=int)),
-    ...         function,
-    ...         symmetric_distance(),
-    ...         symmetric_distance(),
-    ...         stability_map,
-    ...     )
-    
-The resulting Transformation may be used interchangeably with those constructed via the library:
-
-.. doctest::
-
-    >>> from opendp.transformations import *
-    >>> from opendp.measurements import then_base_discrete_laplace
-    >>> trans = (
-    ...     (vector_domain(atom_domain(T=str)), symmetric_distance())
-    ...     >> then_cast_default(TOA=int)
-    ...     >> make_repeat(2)  # our custom transformation
-    ...     >> then_clamp((1, 2))
-    ...     >> then_sum()
-    ...     >> then_base_discrete_laplace(1.0)
-    ... )
-    ...
-    >>> release = trans(["0", "1", "2", "3"])
-    >>> trans.map(1) # computes epsilon
-    4.0
-
-The same holds for measurements and postprocessors.
-You can even mix computational primitives from other DP libraries!
