+++ docs/source/user/transformations.rst
@@ -538,3 +538,66 @@ See the following notebook for more information:
    transformations/aggregation-quantile
 
 These use :func:`opendp.transformations.make_b_ary_tree`, :func:`opendp.transformations.make_consistent_b_ary_tree` and :func:`opendp.transformations.make_quantiles_from_counts`.
+
+
+Bring Your Own
+--------------
+
+Use :func:`opendp.transformations.make_user_transformation` to construct your own transformation.
+
+.. note::
+
+    This requires a looser trust model, as we cannot verify any privacy or stability properties of user-defined functions.
+
+    .. doctest::
+
+        >>> import opendp.prelude as dp
+        >>> dp.enable_features("honest-but-curious")
+
+In this example, we mock the typical API of the OpenDP library to make a transformation that duplicates each record `multiplicity` times:
+
+.. doctest::
+
+    >>> import opendp.prelude as dp
+    >>> from typing import List
+    ...
+    >>> def make_repeat(multiplicity):
+    ...     """Constructs a Transformation that duplicates each record `multiplicity` times"""
+    ...     def function(arg: List[int]) -> List[int]:
+    ...         return arg * multiplicity
+    ... 
+    ...     def stability_map(d_in: int) -> int:
+    ...         # if a user could influence at most `d_in` records before, 
+    ...         # they can now influence `d_in` * `multiplicity` records
+    ...         return d_in * multiplicity
+    ...
+    ...     return dp.t.make_user_transformation(
+    ...         input_domain=dp.vector_domain(dp.atom_domain(T=int)),
+    ...         input_metric=dp.symmetric_distance(),
+    ...         output_domain=dp.vector_domain(dp.atom_domain(T=int)),
+    ...         output_metric=dp.symmetric_distance(),
+    ...         function=function,
+    ...         stability_map=stability_map,
+    ...     )
+    
+The resulting Transformation may be used interchangeably with those constructed via the library:
+
+.. doctest::
+
+    >>> trans = (
+    ...     (dp.vector_domain(dp.atom_domain(T=str)), dp.symmetric_distance())
+    ...     >> dp.t.then_cast_default(TOA=int)
+    ...     >> make_repeat(2)  # our custom transformation
+    ...     >> dp.t.then_clamp((1, 2))
+    ...     >> dp.t.then_sum()
+    ...     >> dp.m.then_base_discrete_laplace(1.0)
+    ... )
+    ...
+    >>> release = trans(["0", "1", "2", "3"])
+    >>> trans.map(1) # computes epsilon
+    4.0
+
+The code snip may form a basis for you to create your own data transformations, 
+and mix them into an OpenDP analysis.
+
+You can also define your own measurements (:func:`opendp.measurements.make_user_measurement`) and postprocessors (:func:`opendp.core.new_function`).
