+++ python/src/opendp/prelude.py
@@ -7,3 +7,4 @@ from opendp.metrics import *
 from opendp.measures import *
 from opendp.typing import *
 from opendp.accuracy import *
+from opendp.core import new_function
