+++ rust/src/combinators/mod.rs
@@ -23,11 +23,6 @@ mod fix_delta;
 #[cfg(feature = "contrib")]
 pub use crate::combinators::fix_delta::*;
 
-#[cfg(feature = "contrib")]
-mod user_defined;
-#[cfg(feature = "contrib")]
-pub use crate::combinators::user_defined::*;
-
 #[cfg(test)]
 pub mod tests {
     use crate::core::{Function, Measurement, PrivacyMap, StabilityMap, Transformation};
