--- rust/src/combinators/user_defined/mod.rs
@@ -1,2 +1,0 @@
-#[cfg(feature = "ffi")]
-mod ffi;
