+++ rust/src/ffi/any.rs
@@ -11,14 +11,24 @@ use std::fmt::{Debug, Formatter};
 
 use crate::core::{
     Domain, Function, Measure, Measurement, Metric, MetricSpace, PrivacyMap, StabilityMap,
-    Transformation,
+    Transformation, FfiResult,
 };
 use crate::error::*;
 use crate::interactive::{Answer, Query, Queryable};
 use crate::{err, fallible};
 
 use super::glue::Glue;
-use super::util::Type;
+use super::util::{Type, into_owned};
+
+
+pub type CallbackFn = extern "C" fn(*const AnyObject) -> *mut FfiResult<*mut AnyObject>;
+
+// wrap a CallbackFn in a closure, so that it can be used in transformations and measurements
+pub fn wrap_func(func: CallbackFn) -> impl Fn(&AnyObject) -> Fallible<AnyObject> {
+    move |arg: &AnyObject| -> Fallible<AnyObject> {
+        into_owned(func(arg as *const AnyObject))?.into()
+    }
+}
 
 /// A trait for something that can be downcast to a concrete type.
 pub trait Downcast {
