+++ rust/src/measurements/mod.rs
@@ -12,6 +12,11 @@ mod laplace;
 #[cfg(feature = "contrib")]
 pub use laplace::*;
 
+#[cfg(all(feature = "honest-but-curious", feature = "ffi"))]
+mod make_user_measurement;
+#[cfg(all(feature = "honest-but-curious", feature = "ffi"))]
+pub use crate::measurements::make_user_measurement::*;
+
 #[cfg(all(feature = "floating-point", feature = "contrib"))]
 mod laplace_threshold;
 #[cfg(all(feature = "floating-point", feature = "contrib"))]
